ostosten lisäys muokkauksia

+++ budjettisovellus/src/entities/purchase.py
@@ -1,5 +1,5 @@
 class Purchase:
-    def __init__(self, purchase: str, price: float, category: str, username: str):
+    def __init__(self, purchase: str, price: int, category: str, username: str):
         self.purchase = purchase
         self.price = price
         self.category = category
